requirements.txt
jax>=0.4.30
jaxlib>=0.4.30
langgraph>=0.2.0
cohere>=5.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
torch>=2.2.0
transformers>=4.40.0
bitsandbytes>=0.43.0
accelerate>=0.30.0
chromadb>=0.5.0
